Add: update comment. Api call success

+++ src/main/java/core/fu4sbackend/repository/CommentRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface CommentRepository extends JpaRepository<Comment, Long> {
+public interface CommentRepository extends JpaRepository<Comment, Integer> {
     @Query(value = "select comment from Comment comment where comment.post.id = ?1")
-    public List<Comment> findByPostId(long id);
+    public List<Comment> findByPostId(Integer id);
 }
